refactor: parse inputs to processInputs
